refactor: restore backend package quality gates
