+++ pkg/services/factory_runtime/internal/services/orchestration/runtime/dispatch_worker_sessions_cutover_test.go
@@ -32,28 +32,6 @@ type blockingAssociationLedger struct {
 	release        chan struct{}
 }
 
-type workerSessionWatermarkedLedger struct {
-	*recordingfixtures.ScriptedRuntimeLedger
-	watermark        recordings.CanonicalEventCursor
-	available        bool
-	calls            int
-	generations      []string
-	watermarkForCall func(int, string) (recordings.CanonicalEventCursor, bool)
-}
-
-func (ledger *workerSessionWatermarkedLedger) CompletedFlushWatermark(
-	streamGenerationID string,
-) (recordings.CanonicalEventCursor, bool) {
-	ledger.calls++
-	ledger.generations = append(ledger.generations, streamGenerationID)
-	if ledger.watermarkForCall != nil {
-		return ledger.watermarkForCall(ledger.calls, streamGenerationID)
-	}
-	return ledger.watermark, ledger.available
-}
-
-var _ recordings.CompletedFlushWatermarkReader = (*workerSessionWatermarkedLedger)(nil)
-
 func newBlockingAssociationLedger() *blockingAssociationLedger {
 	return &blockingAssociationLedger{
 		ScriptedRuntimeLedger: &recordingfixtures.ScriptedRuntimeLedger{},
@@ -329,258 +307,6 @@ func TestRecordedWorkerSessionObservation_ListsHistoricalAttemptsInChronological
 	}
 	if result.Observations[1].AttemptID != "dispatch-late" || result.Observations[1].TurnID != "turn-late" {
 		t.Fatalf("late recorded observation = %#v, want dispatch-late/turn-late", result.Observations[1])
-	}
-}
-
-func TestRecordedWorkerSessionObservationConfirmationFollowsCompletedFlushWatermark(t *testing.T) {
-	base := time.Date(2026, 8, 23, 12, 0, 0, 0, time.UTC)
-	const (
-		generationID    = "generation-worker-confirmation"
-		workID          = "work-worker-confirmation"
-		dispatchID      = "dispatch-worker-confirmation"
-		workerSessionID = "worker-worker-confirmation"
-	)
-	events := []interfaces.FactoryEvent{
-		{
-			Context: interfaces.FactoryEventContext{
-				Tick: 1, Sequence: 1, EventTime: base,
-				DispatchID: stringPointerForRecordedTest(dispatchID),
-				WorkIDs:    stringSliceForRecordedTest([]string{workID}),
-			},
-			Id:   "worker-confirmation-request",
-			Type: interfaces.FactoryEventTypeDispatchRequest,
-		},
-		{
-			Context: interfaces.FactoryEventContext{
-				Tick: 1, Sequence: 2, EventTime: base.Add(time.Second),
-				DispatchID: stringPointerForRecordedTest(dispatchID),
-				RequestID:  stringPointerForRecordedTest("turn-worker-confirmation"),
-			},
-			Id:      "worker-confirmation-association",
-			Type:    interfaces.FactoryEventTypeDispatchWorkerSessionAssoc,
-			Payload: mustMarshalRecordedTest(t, interfaces.DispatchWorkerSessionAssociationEventPayload{WorkerSessionID: workerSessionID}),
-		},
-		{
-			Context: interfaces.FactoryEventContext{
-				Tick: 2, Sequence: 3, EventTime: base.Add(2 * time.Second),
-				DispatchID: stringPointerForRecordedTest(dispatchID),
-			},
-			Id:   "worker-confirmation-response",
-			Type: interfaces.FactoryEventTypeDispatchResponse,
-		},
-	}
-	ledger := &workerSessionWatermarkedLedger{
-		ScriptedRuntimeLedger: &recordingfixtures.ScriptedRuntimeLedger{
-			Events:       events,
-			GenerationID: generationID,
-		},
-		watermark: recordings.CanonicalEventCursor{
-			StreamGenerationID: generationID,
-			Sequence:           3,
-		},
-	}
-	projector := func(_ []interfaces.FactoryEvent, _ int) (interfaces.FactoryWorldState, error) {
-		return interfaces.FactoryWorldState{
-			CompletedDispatches: []interfaces.FactoryWorldDispatchCompletion{{
-				DispatchID: dispatchID, StartedAt: base, CompletedAt: base.Add(2 * time.Second),
-				WorkItemIDs: []string{workID},
-				Result:      interfaces.WorkstationResult{Outcome: string(workers.OutcomeAccepted)},
-			}},
-		}, nil
-	}
-	service := newRecordedWorkerSessionObservation(
-		nil,
-		ledger,
-		factory.WorldStateProjector(projector),
-		platformclock.Real{},
-		nil,
-	)
-
-	listBeforeFlush, err := service.ListObservations(context.Background(), workersessions.ListObservationsRequest{WorkID: workID})
-	if err != nil {
-		t.Fatalf("ListObservations(before flush) error = %v", err)
-	}
-	if len(listBeforeFlush.Observations) != 1 {
-		t.Fatalf("ListObservations(before flush) count = %d, want 1", len(listBeforeFlush.Observations))
-	}
-	before := listBeforeFlush.Observations[0]
-	if before.ConfirmationState != workersessions.ConfirmationStateUnconfirmed || before.StateSequence != 3 || !before.StateSequenceKnown || before.StreamGenerationID != generationID {
-		t.Fatalf("list before flush = %#v, want cursor at sequence 3 and UNCONFIRMED", before)
-	}
-
-	showBeforeFlush, err := service.GetObservationByWorkerSessionID(context.Background(), workersessions.GetObservationByWorkerSessionIDRequest{WorkerSessionID: workerSessionID})
-	if err != nil {
-		t.Fatalf("GetObservationByWorkerSessionID(before flush) error = %v", err)
-	}
-	if showBeforeFlush.ConfirmationState != workersessions.ConfirmationStateUnconfirmed || showBeforeFlush.State != before.State {
-		t.Fatalf("show before flush = %#v, want UNCONFIRMED and same state as list", showBeforeFlush)
-	}
-
-	ledger.available = true
-	listAfterFlush, err := service.ListObservations(context.Background(), workersessions.ListObservationsRequest{WorkID: workID})
-	if err != nil {
-		t.Fatalf("ListObservations(after flush) error = %v", err)
-	}
-	showAfterFlush, err := service.GetObservationByWorkerSessionID(context.Background(), workersessions.GetObservationByWorkerSessionIDRequest{WorkerSessionID: workerSessionID})
-	if err != nil {
-		t.Fatalf("GetObservationByWorkerSessionID(after flush) error = %v", err)
-	}
-	if listAfterFlush.Observations[0].ConfirmationState != workersessions.ConfirmationStateConfirmed || showAfterFlush.ConfirmationState != workersessions.ConfirmationStateConfirmed {
-		t.Fatalf("after flush confirmation = list:%q show:%q, want CONFIRMED for both surfaces", listAfterFlush.Observations[0].ConfirmationState, showAfterFlush.ConfirmationState)
-	}
-	if ledger.calls != 4 || len(ledger.generations) != ledger.calls {
-		t.Fatalf("watermark calls = %d/%v, want one generation-scoped sample per list/show", ledger.calls, ledger.generations)
-	}
-	for _, gotGeneration := range ledger.generations {
-		if gotGeneration != generationID {
-			t.Fatalf("watermark generation = %q, want %q", gotGeneration, generationID)
-		}
-	}
-}
-
-func TestRecordedWorkerSessionObservationListSamplesWatermarkOnceForMergedSources(t *testing.T) {
-	base := time.Date(2026, 8, 24, 12, 0, 0, 0, time.UTC)
-	const (
-		generationID = "generation-worker-list-sample"
-		workID       = "work-worker-list-sample"
-	)
-	events := recordedObservationTestEvents(t, base, workID)
-	ledger := &workerSessionWatermarkedLedger{
-		ScriptedRuntimeLedger: &recordingfixtures.ScriptedRuntimeLedger{
-			Events:       events,
-			GenerationID: generationID,
-		},
-		watermarkForCall: func(call int, generation string) (recordings.CanonicalEventCursor, bool) {
-			if call == 1 {
-				return recordings.CanonicalEventCursor{StreamGenerationID: generation, Sequence: 2}, true
-			}
-			return recordings.CanonicalEventCursor{StreamGenerationID: generation, Sequence: 0}, true
-		},
-	}
-	live := &processLocalWorkerSessionService{
-		observationListResult: workersessions.ListObservationsResult{Observations: []workersessions.Observation{{
-			WorkerSessionID:    "worker-live-only",
-			AttemptID:          "dispatch-live-only",
-			ConfirmationState:  workersessions.ConfirmationStateConfirmed,
-			StreamGenerationID: generationID,
-			StateSequence:      2,
-			StateSequenceKnown: true,
-		}}},
-	}
-	service := newRecordedWorkerSessionObservation(
-		live,
-		ledger,
-		func(_ []interfaces.FactoryEvent, _ int) (interfaces.FactoryWorldState, error) {
-			return interfaces.FactoryWorldState{
-				WorkItemsByID: map[string]work.FactoryWorkItem{workID: {ID: workID}},
-			}, nil
-		},
-		platformclock.Real{},
-		nil,
-	)
-
-	result, err := service.ListObservations(context.Background(), workersessions.ListObservationsRequest{WorkID: workID})
-	if err != nil {
-		t.Fatalf("ListObservations() error = %v", err)
-	}
-	if ledger.calls != 1 || len(ledger.generations) != 1 || ledger.generations[0] != generationID {
-		t.Fatalf("watermark samples = %d/%v, want one sample for %q", ledger.calls, ledger.generations, generationID)
-	}
-	if len(result.Observations) != 3 {
-		t.Fatalf("ListObservations() returned %d observations, want recorded and live observations", len(result.Observations))
-	}
-	for _, observation := range result.Observations {
-		if observation.ConfirmationState != workersessions.ConfirmationStateConfirmed {
-			t.Fatalf("observation %q confirmation = %q, want CONFIRMED from the single sample", observation.WorkerSessionID, observation.ConfirmationState)
-		}
-	}
-}
-
-func TestProcessLocalWorkerSessionObservationsRemainUnconfirmed(t *testing.T) {
-	live := &processLocalWorkerSessionService{
-		topLevelResult: workersessions.ListWorkerSessionObservationsResult{Observations: []workersessions.Observation{{
-			WorkerSessionID:   "worker-live-running",
-			AttemptID:         "attempt-live-running",
-			State:             workersessions.StateRunning,
-			ConfirmationState: workersessions.ConfirmationStateConfirmed,
-		}}},
-		getByWorkerResult: workersessions.Observation{
-			WorkerSessionID:   "worker-live-terminal",
-			AttemptID:         "attempt-live-terminal",
-			State:             workersessions.StateTerminated,
-			ConfirmationState: workersessions.ConfirmationStateConfirmed,
-		},
-	}
-	service := newRecordedWorkerSessionObservation(live, nil, nil, platformclock.Real{}, nil)
-
-	listed, err := service.ListWorkerSessionObservations(context.Background(), workersessions.ListWorkerSessionObservationsRequest{})
-	if err != nil {
-		t.Fatalf("ListWorkerSessionObservations() error = %v", err)
-	}
-	if len(listed.Observations) != 1 || listed.Observations[0].ConfirmationState != workersessions.ConfirmationStateUnconfirmed {
-		t.Fatalf("live list = %#v, want one UNCONFIRMED running observation", listed.Observations)
-	}
-	shown, err := service.GetObservationByWorkerSessionID(context.Background(), workersessions.GetObservationByWorkerSessionIDRequest{WorkerSessionID: "worker-live-terminal"})
-	if err != nil {
-		t.Fatalf("GetObservationByWorkerSessionID() error = %v", err)
-	}
-	if shown.State != workersessions.StateTerminated || shown.ConfirmationState != workersessions.ConfirmationStateUnconfirmed {
-		t.Fatalf("live show = %#v, want terminated UNCONFIRMED observation", shown)
-	}
-}
-
-func TestRecordedWorkerSessionObservationReplayUsesWatermarkForConfirmation(t *testing.T) {
-	base := time.Date(2026, 8, 23, 13, 0, 0, 0, time.UTC)
-	const (
-		generationID    = "generation-worker-replay-confirmation"
-		dispatchID      = "dispatch-worker-replay-confirmation"
-		workerSessionID = "worker-worker-replay-confirmation"
-	)
-	replayEvents := []interfaces.FactoryEvent{
-		{
-			Context: interfaces.FactoryEventContext{Tick: 1, Sequence: 1, EventTime: base, DispatchID: stringPointerForRecordedTest(dispatchID)},
-			Id:      "worker-replay-request",
-			Type:    interfaces.FactoryEventTypeDispatchRequest,
-		},
-		{
-			Context: interfaces.FactoryEventContext{Tick: 1, Sequence: 2, EventTime: base.Add(time.Second), DispatchID: stringPointerForRecordedTest(dispatchID)},
-			Id:      "worker-replay-association",
-			Type:    interfaces.FactoryEventTypeDispatchWorkerSessionAssoc,
-			Payload: mustMarshalRecordedTest(t, interfaces.DispatchWorkerSessionAssociationEventPayload{WorkerSessionID: workerSessionID}),
-		},
-		{
-			Context: interfaces.FactoryEventContext{Tick: 2, Sequence: 3, EventTime: base.Add(2 * time.Second), DispatchID: stringPointerForRecordedTest(dispatchID)},
-			Id:      "worker-replay-response",
-			Type:    interfaces.FactoryEventTypeDispatchResponse,
-		},
-	}
-	ledger := &workerSessionWatermarkedLedger{
-		ScriptedRuntimeLedger: &recordingfixtures.ScriptedRuntimeLedger{GenerationID: generationID},
-		watermark:             recordings.CanonicalEventCursor{StreamGenerationID: generationID, Sequence: 3},
-		available:             true,
-	}
-	service := newRecordedWorkerSessionObservationWithRecording(
-		nil,
-		ledger,
-		func(_ []interfaces.FactoryEvent, _ int) (interfaces.FactoryWorldState, error) {
-			return interfaces.FactoryWorldState{CompletedDispatches: []interfaces.FactoryWorldDispatchCompletion{{
-				DispatchID: dispatchID, StartedAt: base, CompletedAt: base.Add(2 * time.Second),
-				Result: interfaces.WorkstationResult{Outcome: string(workers.OutcomeAccepted)},
-			}}}, nil
-		},
-		platformclock.Real{},
-		nil,
-		replayEvents,
-		"",
-		nil,
-	)
-
-	show, err := service.GetObservationByWorkerSessionID(context.Background(), workersessions.GetObservationByWorkerSessionIDRequest{WorkerSessionID: workerSessionID})
-	if err != nil {
-		t.Fatalf("GetObservationByWorkerSessionID(replay) error = %v", err)
-	}
-	if show.ConfirmationState != workersessions.ConfirmationStateConfirmed || show.StateSequence != 3 || show.StreamGenerationID != generationID {
-		t.Fatalf("replay observation = %#v, want CONFIRMED at generation sequence 3", show)
 	}
 }
