+++ pkg/services/work/transports/cli/list_request_test.go
@@ -7,6 +7,7 @@ import (
 	"testing"
 
 	workdomain "github.com/portpowered/infinite-you/pkg/services/work"
+	factoryapi "github.com/portpowered/infinite-you/pkg/transports/http/generated"
 )
 
 func TestBuildListRequest_NormalizesQueryWithoutHTTP(t *testing.T) {
@@ -63,4 +64,23 @@ func TestBuildListRequest_NormalizesQueryWithoutHTTP(t *testing.T) {
 
 func encodeCursor(value string) string {
 	return base64.StdEncoding.EncodeToString([]byte(value))
+}
+
+func TestNormalizeWorkConfirmationStateDefaultsUnknownValues(t *testing.T) {
+	unknown := factoryapi.ConfirmationState("LEGACY")
+	work := &factoryapi.Work{
+		ConfirmationState: &unknown,
+		StopSummary: &factoryapi.FactoryStopSummary{
+			LatestDispatch: &factoryapi.FactoryStopDispatchSummary{ConfirmationState: unknown},
+		},
+	}
+
+	normalizeWorkConfirmationState(work)
+
+	if work.ConfirmationState == nil || *work.ConfirmationState != factoryapi.UNCONFIRMED {
+		t.Fatalf("work confirmationState = %v, want UNCONFIRMED", work.ConfirmationState)
+	}
+	if work.StopSummary.LatestDispatch.ConfirmationState != factoryapi.UNCONFIRMED {
+		t.Fatalf("dispatch confirmationState = %q, want UNCONFIRMED", work.StopSummary.LatestDispatch.ConfirmationState)
+	}
 }
