+++ pkg/transports/mapping/factorysession/stop_summary_test.go
@@ -56,3 +56,71 @@ func TestStopSummaryToAPIPreservesStructuredSchemaViolationReason(t *testing.T) 
 		t.Fatalf("mapped failure reason = %q, want structured_output_schema_violation", mapped.LatestDispatch.FailureDetail.Reason)
 	}
 }
+
+func TestDispatchReadMappingDefaultsLiveFactsToUnconfirmedAndPreservesConfirmedFacts(t *testing.T) {
+	t.Parallel()
+
+	live := factorysessionmapping.ListDispatchesResponseToAPI(factorysessions.ListDispatchesResult{
+		SessionID: "dur-sess-dispatch-confirmation-001",
+		Dispatches: []factorysessions.DispatchSummary{{
+			ID: "dispatch-live", Status: factorysessions.DispatchStatus("RUNNING"),
+			DispatchKind: "JAVASCRIPT_AGENT",
+		}},
+	})
+	if len(live.Dispatches) != 1 || live.Dispatches[0].ConfirmationState != factoryapi.UNCONFIRMED {
+		t.Fatalf("live dispatch = %#v, want UNCONFIRMED", live.Dispatches)
+	}
+
+	confirmed := factorysessionmapping.ListDispatchesResponseToAPI(factorysessions.ListDispatchesResult{
+		SessionID: "dur-sess-dispatch-confirmation-001",
+		Dispatches: []factorysessions.DispatchSummary{{
+			ID: "dispatch-recorded", Status: factorysessions.DispatchStatus("COMPLETED"),
+			DispatchKind:      "JAVASCRIPT_AGENT",
+			ConfirmationState: factorysessions.ConfirmationStateConfirmed,
+		}},
+	})
+	if len(confirmed.Dispatches) != 1 || confirmed.Dispatches[0].ConfirmationState != factoryapi.CONFIRMED {
+		t.Fatalf("confirmed dispatch = %#v, want CONFIRMED", confirmed.Dispatches)
+	}
+
+	detail := factorysessionmapping.DispatchDetailResponseToAPI(factorysessions.DispatchDetail{
+		DispatchSummary: factorysessions.DispatchSummary{
+			ID: "dispatch-live", Status: factorysessions.DispatchStatus("RUNNING"),
+			DispatchKind: "JAVASCRIPT_AGENT",
+		},
+		SessionID: "dur-sess-dispatch-confirmation-001", OrchestratorKind: "JAVASCRIPT",
+	})
+	if detail.ConfirmationState != factoryapi.UNCONFIRMED {
+		t.Fatalf("live detail confirmationState = %q, want UNCONFIRMED", detail.ConfirmationState)
+	}
+}
+
+func TestHistoricalDispatchReadMappingRequiresExplicitConfirmation(t *testing.T) {
+	t.Parallel()
+
+	input := factorysessionmapping.HistoricalDispatchInput{
+		ID: "dispatch-recorded", Status: "INTERRUPTED", DispatchKind: "PETRI_TRANSITION",
+	}
+	response := factorysessionmapping.HistoricalDispatchListToAPI("dur-sess-dispatch-confirmation-002", []factorysessionmapping.HistoricalDispatchInput{input})
+	if len(response.Dispatches) != 1 || response.Dispatches[0].ConfirmationState != factoryapi.UNCONFIRMED {
+		t.Fatalf("legacy historical dispatch = %#v, want UNCONFIRMED", response.Dispatches)
+	}
+	detail := factorysessionmapping.HistoricalDispatchDetailToAPI(
+		"dur-sess-dispatch-confirmation-002", input, "PETRI",
+	)
+	if detail.ConfirmationState != factoryapi.UNCONFIRMED {
+		t.Fatalf("legacy historical dispatch detail = %q, want UNCONFIRMED", detail.ConfirmationState)
+	}
+
+	input.ConfirmationState = "CONFIRMED"
+	response = factorysessionmapping.HistoricalDispatchListToAPI("dur-sess-dispatch-confirmation-002", []factorysessionmapping.HistoricalDispatchInput{input})
+	if len(response.Dispatches) != 1 || response.Dispatches[0].ConfirmationState != factoryapi.CONFIRMED {
+		t.Fatalf("explicit historical dispatch = %#v, want CONFIRMED", response.Dispatches)
+	}
+	detail = factorysessionmapping.HistoricalDispatchDetailToAPI(
+		"dur-sess-dispatch-confirmation-002", input, "PETRI",
+	)
+	if detail.ConfirmationState != factoryapi.CONFIRMED {
+		t.Fatalf("explicit historical dispatch detail = %q, want CONFIRMED", detail.ConfirmationState)
+	}
+}
